Move test providers to one package
