+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/smoke/WorkbenchRuntimeSmartRouterTwoKieServersTwoDatabasesScenarioIntegrationTest.java
@@ -23,9 +23,9 @@ import org.kie.cloud.api.scenario.WorkbenchRuntimeSmartRouterTwoKieServersTwoDat
 import org.kie.cloud.integrationtests.AbstractCloudIntegrationTest;
 import org.kie.cloud.integrationtests.category.JBPMOnly;
 import org.kie.cloud.integrationtests.category.Smoke;
-import org.kie.cloud.integrationtests.smoke.testproviders.FireRulesTestProvider;
-import org.kie.cloud.integrationtests.smoke.testproviders.OptaplannerTestProvider;
-import org.kie.cloud.integrationtests.smoke.testproviders.ProcessTestProvider;
+import org.kie.cloud.integrationtests.testproviders.FireRulesTestProvider;
+import org.kie.cloud.integrationtests.testproviders.OptaplannerTestProvider;
+import org.kie.cloud.integrationtests.testproviders.ProcessTestProvider;
 import org.kie.cloud.integrationtests.util.ScenarioDeployer;
 import org.kie.cloud.maven.constants.MavenConstants;
