+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/sso/AuthSsoIntegrationTest.java
@@ -27,8 +27,8 @@ import org.kie.cloud.api.deployment.constants.DeploymentConstants;
 import org.kie.cloud.api.scenario.WorkbenchKieServerScenario;
 import org.kie.cloud.integrationtests.AbstractCloudIntegrationTest;
 import org.kie.cloud.integrationtests.smoke.WorkbenchKieServerPersistentScenarioIntegrationTest;
-import org.kie.cloud.integrationtests.smoke.testproviders.ProcessTestProvider;
-import org.kie.cloud.integrationtests.sso.testproviders.PersistenceTestProvider;
+import org.kie.cloud.integrationtests.testproviders.ProcessTestProvider;
+import org.kie.cloud.integrationtests.testproviders.PersistenceTestProvider;
 import org.kie.cloud.integrationtests.util.ScenarioDeployer;
 import org.kie.cloud.integrationtests.util.WorkbenchUtils;
 import org.kie.cloud.maven.constants.MavenConstants;
