+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/testproviders/FireRulesTestProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.cloud.integrationtests.smoke.testproviders;
+package org.kie.cloud.integrationtests.testproviders;
 
 import static org.assertj.core.api.Assertions.assertThat;
