+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/testproviders/OptaplannerTestProvider.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.cloud.integrationtests.smoke.testproviders;
+package org.kie.cloud.integrationtests.testproviders;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
