+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/testproviders/PersistenceTestProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.cloud.integrationtests.sso.testproviders;
+package org.kie.cloud.integrationtests.testproviders;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import org.kie.cloud.api.deployment.Deployment;
